Fix a number of CLANG warnings

+++ base/random.h
@@ -28,7 +28,7 @@ extern "C" {
 
 #include <base/types.h>
 
-#define _aax_random()		((double)xoroshiro128plus()/UINT64_MAX)
+#define _aax_random()	((double)xoroshiro128plus()/(double)UINT64_MAX)
 
 void _aax_srandom();
 uint64_t xorshift128plus();
