+++ src/backends/linux/pulseaudio.c
@@ -1209,7 +1209,7 @@ _pulseaudio_set_volume(_driver_t *handle, _aaxRingBuffer *rb, ssize_t offset, ui
 
       gain = (float)vol/PA_VOLUME_NORM / fabsf(volume);
 
-      if (fabsf(handle->volume - vol) > 0)
+      if (fabsf((float)handle->volume - vol) > 0)
       {
          pa_cvolume cvol;
